Adds shipping service with Canada Post integration
Implements a shipping service to retrieve shipping quotes.
This commit introduces the `ShippingService` class, which currently integrates with the Canada Post API to generate shipping quotes based on the destination and weight of the package.
It also introduces DTOs and Models for processing the data, and dependency injections using the IConfiguration interface for secrets.
The shipping service determines the shipping type and approximate delivery time.