Adds Purolator shipping quote generation
Implements the ability to generate shipping quotes using the Purolator API.

This involves fetching an OAuth token, building a request XML,
sending it to the Purolator Estimating Service, and parsing the XML
response to extract shipping options and costs. It also includes
mapping Purolator service codes to generic shipping types.